refactor(probability-sampling): move randomness decoding and parsing to dedicated class
